+++ src/main/java/net/neganote/gtutilities/GregTechModernUtilities.java
@@ -25,6 +25,7 @@ import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
 import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 import net.neganote.gtutilities.client.renderer.UtilShaders;
+import net.neganote.gtutilities.common.data.UtilPlaceholders;
 import net.neganote.gtutilities.common.item.UtilItems;
 import net.neganote.gtutilities.common.machine.UtilMachines;
 import net.neganote.gtutilities.common.materials.UtilMaterials;
@@ -94,6 +95,7 @@ public class GregTechModernUtilities {
         UtilItems.init();
         REGISTRATE.registerRegistrate();
         UtilDatagen.init();
+        UtilPlaceholders.init();
     }
 
     public static ResourceLocation id(String path) {