+++ changelog.md
@@ -1,3 +1,6 @@
-# GregTech Modern Utilities version 2.1.1
+# GregTech Modern Utilities version 2.2.0
 
-* Fix crashing when omnibreaker is disabled
+* Update to GTm 7.1.3
+* Add a placeholder for the WAT to see what its current frequency is
+* Update omni-breaker recipe to use 2 non-full amps
+* Update 64A converter recipes to use non-full amps